🧪 [atlas] Final polish and consolidation of Atlas-Diff tests
- Consolidated all new unit and integration tests into the canonical location `merger/lenskit/tests/test_atlas_diff.py`.
- Removed redundant `tests/test_atlas_diff.py`.
- Performed final hygiene by removing unused imports (`Path`, `monkeypatch`, `os` in 일부 scope).
- Ensured consistent docstrings and `encoding="utf-8"` usage.
- Verified all 16 tests in the consolidated file pass successfully.
- Maintained focused coverage on `_compare_file_sets`, `_load_inventory_index`, and `compute_snapshot_comparison`.
